fix: soft delete route (#2256)
| Q            | A
|--------------| ---
| Bug fix?     | yes
| New feature? | no
| BC-Break?    | no
| License      | MIT

If a resource with a route was deleted, then the route still exists but refer to a null object, because doctrine will filter it out. This leads to unexpected behaviour if a new route was created with the same path. So if a route, which is soft deleted will get invisible by overwriting the staticPrefix with random data and setting the condition to false.